requirements.txt
pyserial>=3.5
pyyaml>=6.0
pytest>=7.0
pytest-mock>=3.10
